make code nicer, template is class
